requirements.txt
# List all python libraries for the lambda
tensorflow==2.4.0
tensorflow_hub==0.11
pillow==8.4.0
humanfriendly==10.0
matplotlib==3.4.3
tqdm==4.62.3
jsonpickle==2.0.0
statistics==1.0.3.5
requests==2.26.0